Create areaCalc.py
Python is especially useful for doing math and can be used to automate many calculations. In this project, we’ll create a calculator that can compute the area of the following shapes:

Circle
Triangle
The program should do the following:

Prompt the user to select a shape.
Calculate the area of that shape.
Print the area of that shape to the user.